rename files to be capitalized
