+++ src/index.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 import './index.css'
 
-import App from './components/app/app'
+import App from './components/app/App'
 
 console.log('Starting Application')
